refactor(shared): move engine-motive and the face reader where a browser can use them
Three files in the video pipeline imported `isEngineAuthoredMotive` from the
server barrel, which drags better-sqlite3, discord.js and ollama in behind it.
That made the whole normalizer layer unbundleable for the browser — the exact
hazard the comment on live-frame.types.ts warns about.

The prefixes and REPETITION_GUARD_MARKER are plain constants with no runtime
dependency, so they move to shared and the old server paths re-export them.
Nothing else changes an import.

faceFor and emotionLabel move too, for a different reason: the web stage has to
resolve a recorded emotion into an expression, and two implementations of that
would drift. avatar.ts keeps its own palette — the film is dark and purple and
the web stage is not — but no longer owns the reading. RecordedEmotion follows
so both sides describe the same shape.
