+++ packages/eval/src/video/avatar.ts
@@ -1,4 +1,10 @@
-import type { RecordedEmotion } from './types.js';
+/**
+ * The MP4 renderer's figure. Its palette is its own — the film is dark and
+ * purple, the web stage is paper and highlighter — but the face logic is not:
+ * `faceFor` and `emotionLabel` live in `@perfectman/shared` so both renderers
+ * read a recorded emotion the same way.
+ */
+export { emotionLabel, faceFor } from "@perfectman/shared";
 
 const colors = ['#d8d9dd', '#202022', '#6b55f5', '#746b8b', '#424653', '#827786'];
 export const avatarColor = (index: number): string => colors[index % colors.length]!;
@@ -31,53 +37,3 @@ export function avatar(index: number): string {
     </g></svg>`;
 }
 
-// Only top-level dimensions are eligible. Per-target relational values remain private context.
-const socialFaces: Record<string, string> = {
-  resentment: 'angry', contempt: 'angry', jealousy: 'angry', envy: 'angry',
-  humiliation: 'worried', shame: 'worried', suspicion: 'worried', neediness: 'worried',
-  socialAnxiety: 'worried', fearOfExclusion: 'worried',
-  affection: 'smile', admiration: 'smile', pride: 'smile', joy: 'smile', relief: 'smile',
-};
-function dominantSocial(values: Record<string, number>): [string, number] | undefined {
-  return Object.entries(values).filter(([key, value]) => socialFaces[key] && value >= .6)
-    .sort((a, b) => b[1] - a[1])[0];
-}
-
-/** A visual approximation of recorded state, never a dialogue classifier. */
-export function faceFor(emotion?: RecordedEmotion): string {
-  if (!emotion) return 'neutral';
-  const label = emotion.label?.toLowerCase().trim();
-  const labels: Record<string, string> = {
-    angry: 'angry', anger: 'angry', rage: 'angry', furious: 'angry', irritated: 'angry',
-    worried: 'worried', afraid: 'worried', anxious: 'worried', sad: 'worried',
-    tired: 'tired', bored: 'tired', sleepy: 'tired',
-    smile: 'smile', happy: 'smile', joy: 'smile', relieved: 'smile', satisfied: 'smile',
-    shock: 'shock', shocked: 'shock', surprised: 'shock', neutral: 'neutral', calm: 'neutral',
-  };
-  // Replay labels can be raw dimension names; unrecognized names must still use recorded values.
-  if (label && labels[label]) return labels[label]!;
-  const value = emotion.values ?? {};
-  const drivers = (emotion.drivers ?? []).map(d => d.toLowerCase().replaceAll('_', ''));
-  if (drivers.some(d => ['anger', 'resentment', 'hostility', 'irritation'].includes(d))) return 'angry';
-  if (drivers.some(d => ['fear', 'anxiety', 'socialanxiety', 'shame', 'guilt'].includes(d))) return 'worried';
-  if (drivers.some(d => ['joy', 'relief', 'gratitude', 'warmth', 'affection'].includes(d))) return 'smile';
-  const social = dominantSocial(value);
-  if (social) return socialFaces[social[0]]!;
-  if (typeof value.valence === 'number') {
-    if (value.valence < -0.25) return (value.arousal ?? 0) > 0.65 ? 'angry' : 'worried';
-    if (value.valence > 0.25) return 'smile';
-  }
-  if (typeof value.energy === 'number' && value.energy < 0.25) return 'tired';
-  return 'neutral';
-}
-
-export function emotionLabel(emotion?: RecordedEmotion): string {
-  if (!emotion) return '';
-  const prefix = emotion.source === 'authored' ? 'Authored' : 'Recorded';
-  if (emotion.label) return `${prefix}: ${emotion.label}`;
-  if (emotion.drivers?.length) return `${prefix} drivers: ${emotion.drivers.join(', ')}`;
-  const values = Object.entries(emotion.values ?? {}).filter(([key]) => ['valence', 'arousal', 'energy'].includes(key));
-  const social = dominantSocial(emotion.values ?? {});
-  if (social) values.unshift(social);
-  return values.length ? `${prefix} ${values.map(([key, value]) => `${key} ${value}`).join(' · ')}` : `${prefix} state`;
-}
